refactor: 로고 홈으로 리디렉션 수정,,스크롤바 수정,,관리자페이지 관리자 계정 외 접근불가 수정
