+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { RouterProvider } from 'react-router-dom';
-import { router } from './Routers/Index';
+import { router } from './Routers/index';
 import './App.css';
 
 function App() {
