+++ src/components/slider/PlaylistSlider.tsx
